fix(terminal): detect PR links with ANSI resets (#8126)
